*fix bug in TestDarknet application; change default build location.

+++ check.sh
@@ -11,8 +11,8 @@ DIR=./data/"$FRAMEWORK"/"$NAME"
 PATHES="-fm=$DIR/other.dsc -fw=$DIR/other.dat -sm=$DIR/synet$FORMAT.xml -sw=$DIR/synet$FORMAT.bin -id=$DIR/image -od=$DIR/output -tp=$DIR/param.xml"
 PREFIX="${FRAMEWORK:0:1}"
 OUT=./test/check/"$DATE_TIME$PREFIX"
-LOG="$OUT"/cl"$PREFIX"_"$NAME"_f"$FORMAT"_b"$BATCH".txt
-BIN_DIR=./build_"$FRAMEWORK"
+LOG="$OUT"/c"$PREFIX"_"$NAME"_f"$FORMAT"_b"$BATCH".txt
+BIN_DIR=./build #_"$FRAMEWORK"
 BIN="$BIN_DIR"/test_"$FRAMEWORK"
 
 if [ "${NAME:0:5}" = "test_" ] && [ "${NAME:8:9}" = "i" ]; then
@@ -24,23 +24,15 @@ fi
 
 echo $LOG
 
-if [ -f $DIR/image/descript.ion ];then
-	rm $DIR/image/descript.ion
-fi
+if [ -f $DIR/image/descript.ion ];then rm $DIR/image/descript.ion; fi
 
 export LD_LIBRARY_PATH="$BIN_DIR":$LD_LIBRARY_PATH
 
 "$BIN" -m=convert $PATHES -tf=$FORMAT
-if [ $? -ne 0 ];then
-  echo "Test $DIR is failed!"
-  exit
-fi
+if [ $? -ne 0 ]; then echo "Test $DIR is failed!"; exit; fi
 
 "$BIN" -m=compare -e=3 $PATHES -if=*.* -rn=$NUMBER -wt=1 -tt=$THREAD -tf=$FORMAT -bs=$BATCH -t=$THRESHOLD -cs=1 -ln=$LOG
-if [ $? -ne 0 ];then
-  echo "Test $DIR is failed!"
-  exit
-fi
+if [ $? -ne 0 ]; then echo "Test $DIR is failed!"; exit; fi
 }
 
 function TEST_I {
@@ -51,9 +43,7 @@ NUMBER=1
 THREAD=0
 TEST $FRAMEWORK $NAME $NUMBER $THREAD 0 1
 TEST $FRAMEWORK $NAME $NUMBER $THREAD 1 1
-if [ $TEST_BATCH -eq 1 ];then
-  TEST $FRAMEWORK $NAME $NUMBER $THREAD 1 2
-fi
+if [ $TEST_BATCH -eq 1 ];then TEST $FRAMEWORK $NAME $NUMBER $THREAD 1 2; fi
 }
 
 function TEST_ALL_I {
